flash: Fix a duplicate early check
We were previously trying to check to see if ${PART_IMAGE} was around before we
defined its contents. It was harmless, but silly. This moves the bootloader
flashing below the definition and the check so that we don't accidentally flash
the bootloader before we check to see if we have a gpt.

Change-Id: Ia00222e4b71b3448d26bb846b6d6c67544793ae2